requirements.txt
Django==4.2.16
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==11.2.1
djangorestframework-simplejwt==5.3.1
